+++ db/migrate/20160911182049_add_category_to_accounts.rb
@@ -1,0 +1,5 @@
+class AddCategoryToAccounts < ActiveRecord::Migration[5.0]
+  def change
+    add_column :accounts, :category, :string, null: false, default: 'savings'
+  end
+end